/study complete

+++ server/controllers/study/study.js
@@ -56,7 +56,8 @@ module.exports = {
           study: {
             id,
             username: findUsername.username,
-            content: { title, description: content },
+            title,
+            content,
             kakaoLink,
             closed,
             location: findLocation,
@@ -75,8 +76,19 @@ module.exports = {
   // ! 완료
   post: async (req, res) => {
     try {
-      const { content, kakaoLink, startDate, closed, location, title, language_id, loginMethod } =
-        req.body;
+      const {
+        username,
+        content,
+        kakaoLink,
+        startDate,
+        closed,
+        location,
+        title,
+        language_id,
+        loginMethod,
+      } = req.body;
+
+      console.log(req.body);
 
       const data = checkAccessToken(req);
       const { id } = req.params;
@@ -88,8 +100,28 @@ module.exports = {
         return res.status(401).json("wrong req params");
       }
 
+      if (
+        !username ||
+        !content ||
+        !kakaoLink ||
+        !startDate ||
+        closed === undefined ||
+        !location ||
+        !title ||
+        !language_id ||
+        loginMethod === undefined
+      ) {
+        return res.status(401).json("body required");
+      }
+
       const findOrCreateLocation = await Location.findOrCreate({
-        where: { latitude: location[0], longitude: location[1] },
+        where: {
+          latitude: location[0],
+          longitude: location[1],
+          guType: location[2],
+          dongType: location[3],
+          name: location[4],
+        },
       });
 
       // TODO: 경도 위도 필요시
@@ -125,9 +157,131 @@ module.exports = {
       return res.status(500).json();
     }
   },
-  patch: (req, res) => {
+  patch: async (req, res) => {
     try {
-      res.send("study patch ok");
+      const data = checkAccessToken(req);
+
+      const {
+        study_id,
+        username,
+        content,
+        kakaoLink,
+        startDate,
+        closed,
+        location,
+        title,
+        language_id,
+        loginMethod,
+      } = req.body;
+
+      const { id } = req.params;
+
+      if (!data) {
+        return res.status(401).json("signin required");
+      }
+
+      const studyInfo = await Study.findOne({
+        where: { id },
+        include: [
+          { model: Location, as: "location" },
+          { model: Language, as: "language" },
+        ],
+      });
+
+      let studylangueId = [];
+      for (let i = 0; i < studyInfo.language.length; i++) {
+        studylangueId.push(studyInfo.language[i].Study_language.language_id);
+      }
+
+      await Study.update(
+        {
+          content: content ? content : studyInfo.content,
+          title: title ? title : studyInfo.title,
+          kakaoLink: kakaoLink ? kakaoLink : studyInfo.kakaoLink,
+          startDate: startDate ? startDate : studyInfo.startDate,
+          closed: closed ? true : false,
+          updatedAt: new Date(),
+        },
+        { where: { id } }
+      );
+
+      const study_languageInfo = await Study_language.findAll({
+        where: { study_id: id },
+        attributes: ["id"],
+      });
+      const study_languageId = [];
+
+      study_languageInfo.forEach((el) => study_languageId.push(el.id));
+
+      if (language_id) {
+        if (language_id.length === study_languageId) {
+          for (let i = 0; i < language_id.length; i++) {
+            await Study_language.update(
+              {
+                language_id: language_id[i],
+              },
+              { where: { id: study_languageId[i] } }
+            );
+          }
+        } else {
+          for (let i = 0; i < study_languageId.length; i++) {
+            Study_language.destroy({
+              where: { id: study_languageId[i] },
+            });
+          }
+
+          for (let i = 0; i < language_id.length; i++) {
+            Study_language.create({
+              study_id: id,
+              language_id: language_id[i],
+            });
+          }
+        }
+      }
+
+      if (location) {
+        await Location.update(
+          {
+            name: location[4],
+            latitude: location[0],
+            longitude: location[1],
+            guType: location[2],
+            dongType: location[3],
+          },
+          { where: { id: studyInfo.location_id } }
+        );
+      }
+
+      const result = await Study.findOne({
+        where: { id },
+        include: [
+          { model: Location, as: "location" },
+          { model: Language, as: "language" },
+          { model: User, as: "user" },
+        ],
+      });
+
+      res.status(200).json({
+        data: {
+          studies: [
+            {
+              id: result.id,
+              username: result.user.username,
+              content: result.content,
+              title: result.title,
+              kakaoLink: result.kakaoLink,
+              closed: result.closed,
+              location: {
+                latitude: result.location.latitude,
+                longitude: result.location.longitude,
+              },
+              language: result.language,
+              createdAt: result.createdAt,
+              updatedAt: result.updatedAt,
+            },
+          ],
+        },
+      });
     } catch (err) {
       console.error(err);
       return res.status(500).json();
@@ -136,6 +290,11 @@ module.exports = {
   // ! 완료
   delete: async (req, res) => {
     try {
+      const data = checkAccessToken(req);
+
+      if (!data) {
+        return res.status(401).json("signin required");
+      }
       const { id } = req.params;
       await Study_comment.destroy({
         where: { study_id: id },
